refactor(LottoGame.java): cast long to double earnings and cost 수익률을 계산할 떄 소수자리의 정확한 표기를 위해 long을 double로 타입 변환 하여 계산하였습니다.

+++ src/main/java/model/LottoGame.java
@@ -55,7 +55,7 @@ public class LottoGame {
     }
 
     public double calculateEarningRate(long earnings, long cost) {
-        return earnings / cost;
+        return (double) earnings / (double) cost;
     }
 
     public List<Lotto> getLottos() {
